feat: add NPM option for viewing documentation
Add an NPM based alternative to using Docker Compose for viewing the
OpenAPI documentation.

Swagger UI should be used as the viewer, with webpack for bundling
everything.
Configure webpack to grab the generated the index.html automatically, it
should include the CSS from the `swagger-ui` module as well as the
`openapi.yml` file itself.

`npm run start` should bundle and serve the documentation using webpack.

NO-TICKET