feat: add name to cctCalc and make wte optional
Update `cctCalculation` schema to allow a custom name for the
calculation. This will allow easier identification by the user when
multiple calculations are saved.

Update `cctChange` to make the `endDate` property optional. In the case
of LTFT any changes are for the remaining training duration, allowing a
custom end date suggests that a temporary WTE change can be applied for.

TIS21-6544