refactor: CCT calculation endpoints
Change the API path for CCT calculations from `/forms` to `/trainee`.
The CCT calculations no longer have a form-like lifecycle, given the
potential need of auto-updating PM data in the calculation the best
remaining fit is the trainee-details service where PM data is held.

Split the `cctCalculation` schema in to `cctCalculationSummary` and
`cctCalculationDetail`, the existing `GET: /trainee/cctCalculation`
endpoint will now return an array of `cctCalculationSummary`. `PUT` and
`POST` endpoints will continue to use `cctCalculationDetail`.

Create a new `GET: /trainee/cctCalculation/{id}` to return the full
details of a single CCT calculation.

Change the `cctCalculation.id` property to read-only so that it is no
longer shown in the `POST` and `PUT` request bodies. All responses will
still include the id property.

TIS21-6544