refactor: split OpenAPI description
The API description will get difficult to work with as it gets larger,
it should be split to separate components.
Use redocly CLI to split the openapi.yml file in to a standardised
multi-file structure.

Update the Docker and NPM configurations to bundle the split files back
in to a single openapi.yml file for consumption by Swagger UI.

NO-TICKET