poc: generate java DTOs from the OpenAPI spec
Create a proof of concept to show how Java DTOs can be generated and
published automatically based on the OpenAPI documentation.

The `gradlew publishToMavenLocal` task can be used to publish the DTOs,
which can be imported as a dependency using
`uk.nhs.tis.trainee:contract-dtos:0.0.1`.

Note: The POC is a combination of `gradle init` defaults, copied
publishing code from `tis-trainee-version-catalog` and AI
generation/debugging for generation and linking the tasks. Some
erroneous config/properties have already been removed but nothing in the
Gradle build file should be trusted as accurate.

NO-TICKET